Fix donate buttons
