+++ app/src/main/java/moe/chenxy/extrasettings/MainActivity.java
@@ -1,27 +1,12 @@
 package moe.chenxy.extrasettings;
 
-import android.didikee.donate.*;
-import java.io.*;
-
-import android.graphics.BitmapFactory;
-import android.os.Build;
 import android.os.Bundle;
-
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
-
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
-
-import android.os.Environment;
-import android.view.View;
-import android.view.Menu;
 import butterknife.ButterKnife;
 import butterknife.InjectView;
-import androidx.preference.Preference;
 import android.view.MenuItem;
-import android.view.WindowManager;
 
 public class MainActivity extends AppCompatActivity{
 
@@ -35,26 +20,6 @@ public class MainActivity extends AppCompatActivity{
         ButterKnife.inject(this);
         initToolbar();
         getFragmentManager().beginTransaction().replace(R.id.content_frame, new SettingFragment()).commit();
-        // donate
-        Preference donate_wechat = getPreferenceManager()
-                .findPreference("wechat");
-        donate_wechat.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                donateWeixin();
-                return false;
-            }
-        });
-        Preference donate_alipay = new Preference().getPreferenceManager().findPreference("alipay");
-        donate_alipay.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                donateAlipay("FKX03133TLJFCY8UNXHC56");
-                return false;
-            }
-        });
     }
 
 
@@ -82,31 +47,6 @@ public class MainActivity extends AppCompatActivity{
                 finish();
                 return true;
         }
-        return false;
-    }
-    /**
-     * 支付宝支付
-     *
-     * @param payCode 收款码后面的字符串；例如：收款二维码里面的字符串为 HTTPS://QR.ALIPAY.COM/FKX03133TLJFCY8UNXHC56 ，则
-     *                payCode = FKX03133TLJFCY8UNXHC56
-     *                注：不区分大小写
-     */
-    public void donateAlipay(String payCode) {
-        boolean hasInstalledAlipayClient = AlipayDonate.hasInstalledAlipayClient(this);
-        if (hasInstalledAlipayClient) {
-            AlipayDonate.startAlipayClient(this,payCode);
-        }
-    }
-
-    /**
-     * 需要提前准备好 微信收款码 照片，可通过微信客户端生成
-     *
-     */
-    public void donateWeixin() {
-        InputStream weixinQrIs = getResources().openRawResource(R.raw.weixin);
-        String qrPath = Environment.getExternalStorageDirectory().getAbsolutePath() + File.separator + "AndroidDonateSample" + File.separator +
-                "weixin.png";
-        WeiXinDonate.saveDonateQrImage2SDCard(qrPath, BitmapFactory.decodeStream(weixinQrIs));
-        WeiXinDonate.donateViaWeiXin(this, qrPath);
+        return true;
     }
 }
